Refactor and improve.
