+++ src/components/util/saveAs.ts
@@ -1,0 +1,8 @@
+export default function(blob: Blob, filename: string) {
+    const downloadLink = document.createElement('a');
+    downloadLink.href = URL.createObjectURL(blob);
+    downloadLink.download = filename;
+    document.body.appendChild(downloadLink);
+    downloadLink.click();
+    document.body.removeChild(downloadLink);
+};
